Updated tutorial
