+++ src/index.html
@@ -9,9 +9,9 @@
 
   <body>
     <ul>
-      <li><a href="/sidebar/">Sidebar App</a></li>
-      <li><a href="/popup/">Popup App</a></li>
-      <li><a href="/tutorial/">Tutorial App</a></li>
+      <li><a href="presentation/apps/sidebar/">Sidebar App</a></li>
+      <li><a href="presentation/apps/popup/">Popup App</a></li>
+      <li><a href="presentation/apps/tutorial/">Tutorial App</a></li>
     </ul>
   </body>
 </html>
