+++ src/presentation/apps/tutorial/Tutorial.css
@@ -1,7 +1,6 @@
 .tutorial-popup {
   z-index: 75;
-  background-color: var(--bg-primary);
-  color: var(--text-primary);
+  background: #fff;
   border-radius: 12px;
   width: 400px;
   padding: 0;
@@ -36,8 +35,8 @@
 }
 
 .tutorial-btn {
-  background: #e0e0e0;
-  color: #333;
+  background: #233b86;
+  color: white;
   border: none;
   border-radius: 6px;
   width: 40%;
@@ -49,12 +48,6 @@
   transition: background 0.2s;
 }
 
-.tutorial-btn.primary {
-  background: #233b86;
-  color: var(--text-primary);
-  font-weight: 500;
-}
-
-.tutorial-btn.primary:hover {
+.tutorial-btn:hover {
   background: #1a4c9b;
 }
