+++ src/presentation/apps/tutorial/index.css
@@ -17,10 +17,7 @@
   }
 }
 
+:root,
 body {
-  background-color: var(--bg-primary);
-  color: var(--text-primary);
-  transition:
-    background-color 0.3s,
-    color 0.3s;
+  background: transparent !important;
 }
